fix: close currentStock gaps — explicit save, updateInventory recalc, Long return type
- Add explicit second inventoryRepository.save() after setCurrentStock() in createInventory()
  so currentStock is persisted safely without relying on implicit JPA dirty-checking alone
- Recalculate currentStock in base units when purchaseQuantity changes in updateInventory()
  preventing stale stock values after a purchase quantity correction
- Change Inventory.getCurrentStock() return type from Integer to Long to match the field type,
  eliminating potential overflow for large stock values and the type inconsistency
- Update all callers in InventoryService and ProductService to use Long/intValue() casts
- Update InventoryServiceTest to expect two save() calls in createInventory (initial + currentStock)

+++ backend/src/main/java/com/princely/shopmanager/core/service/ProductService.java
@@ -325,8 +325,8 @@ public class ProductService extends ShopAwareService {
 
         List<Inventory> inventories = inventoryRepository.findByProductId(productId);
 
-        int totalStock = inventories.stream()
-            .mapToInt(Inventory::getCurrentStock)
+        int totalStock = (int) inventories.stream()
+            .mapToLong(Inventory::getCurrentStock)
             .sum();
 
         int reservedStock = inventories.stream()
@@ -437,8 +437,8 @@ public class ProductService extends ShopAwareService {
 
         return allProducts.stream()
             .filter(product -> {
-                int totalStock = inventoryRepository.findByProductId(product.getId()).stream()
-                    .mapToInt(Inventory::getCurrentStock)
+                long totalStock = inventoryRepository.findByProductId(product.getId()).stream()
+                    .mapToLong(Inventory::getCurrentStock)
                     .sum();
                 return totalStock == 0;
             })
